add imports

+++ src/main/java/com/camel/example/MyService.java
@@ -8,7 +8,10 @@ import java.time.LocalDate;
 import java.time.ZoneId;
 import java.time.ZonedDateTime;
 import java.time.temporal.ChronoUnit;
+import java.time.temporal.TemporalAdjusters;
 import java.util.Date;
+import java.util.Locale;
+
 import org.apache.camel.CamelContext;
 import org.apache.camel.Header;
 import org.springframework.stereotype.Service;
@@ -18,7 +21,7 @@ public class MyService {
 
     CamelContext context;
     public int helloWithName(@Header("input") String input) throws ParseException {
-    	SimpleDateFormat simpleDateFormat = new SimpleDateFormat( "yyyy-MM-dd", Locale.ENGLISH);
+    SimpleDateFormat simpleDateFormat = new SimpleDateFormat( "yyyy-MM-dd", Locale.ENGLISH);
     String dateInString ="2021-07-16";
     Date input_date = simpleDateFormat.parse(dateInString); 
     Instant instantj = input_date.toInstant();
@@ -35,7 +38,7 @@ public class MyService {
     LocalDate firstFriday = start.with(TemporalAdjusters.nextOrSame(DayOfWeek.FRIDAY));
     LocalDate lastFriday = end.with(TemporalAdjusters.previousOrSame(DayOfWeek.FRIDAY));
     count_friday = ChronoUnit.WEEKS.between(firstFriday, lastFriday);
-    return count_friday + 1;
+    return (int) (count_friday + 1);
 
      }
 }
